DOC: Update haroldpolyadd function (#36)
* DOC: Update haroldpolyadd function

Updated the haroldpolyadd function. It now trims the zeros of the output
instead the zeros of the input. Improved the docstring with a parameter
and return list.

* DOC: Improve docstrings for polynomial_ops

Improved the docstrings of some functions in the polynomial_ops. They
have now a parameter and returns list. Also, added examples for the
haroldpolyaddfunctions and reverted the changes I made earlier where I
changed the trimming behaviour.

* DOC: Added NumPy links and some more cleanups

* REV: reverted polyadd code back due to test fails.

I'll address the functionality in another PR.